% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{all_lkps_maps_to_db}
\alias{all_lkps_maps_to_db}
\title{Build a SQLite database of clinical code look up and mapping tables}
\usage{
all_lkps_maps_to_db(all_lkps_maps, db_path, overwrite = FALSE)
}
\arguments{
\item{all_lkps_maps}{A named list of look up and mapping tables, as created
by \code{\link{build_all_lkps_maps}}.}

\item{db_path}{Where the database will be created. If an SQLite database file
already exists here, then the lookup and mapping tables will be added to
this. If \code{NULL} (default), then no database will be created/modified.}

\item{overwrite}{If \code{TRUE}, overwrite tables in the database if they
already exist. Default value is \code{FALSE}.}
}
\value{
Returns \code{db_path} invisibly
}
\description{
Write the output from \code{\link{build_all_lkps_maps}} to a SQLite database.
}
