% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{build_all_lkps_maps_dummy}
\alias{build_all_lkps_maps_dummy}
\title{Create a dummy all_lkps_maps}
\usage{
build_all_lkps_maps_dummy()
}
\value{
A named list of tibbles.
}
\description{
A thin convenience wrapper around \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}}, using dummy data
included with this package.
}
\examples{
build_all_lkps_maps_dummy()
}
\seealso{
\code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}}

Other Dummy data: 
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_caliber_dir_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_all_lkps_maps_dummy}()},
\code{\link{read_ukb_codings_dummy}()}
}
\concept{Dummy data}
