% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{check_codes_exist}
\alias{check_codes_exist}
\title{Utility function - check if all of a set of codes are recognised}
\usage{
check_codes_exist(
  codes,
  lkp_codes,
  code_type,
  return_unrecognised_codes = FALSE
)
}
\arguments{
\item{codes}{Codes being checked}

\item{lkp_codes}{Character vector of lookup codes. Any \code{codes} not present in
\code{lkp_codes} are considered to be unrecognised.}

\item{code_type}{String}

\item{return_unrecognised_codes}{If \code{TRUE}, return a character vector of
unrecognised codes. Default is \code{FALSE}.}
}
\value{
Return vector of unrecognised codes if \code{return_missing_codes} is
\code{TRUE}, otherwise called for side effect (error if any unrecognised codes)
}
\description{
Utility function - check if all of a set of codes are recognised
}
