% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{child_codes_sct}
\alias{child_codes_sct}
\title{Get child codes for one or more SNOMED codes}
\usage{
child_codes_sct(
  conceptIds,
  include_self = TRUE,
  standardise_output = TRUE,
  limit = 1000
)
}
\arguments{
\item{conceptIds}{a character vector of concept identifiers}

\item{include_self}{a logical vector indicating whether the \code{conceptIds}
should be included in the results. The default is \code{TRUE}. If a single
value is provided, it will be recycled.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns for the relevant lookup sheet from
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank
resource 592}).}

\item{limit}{Number of SNOMED codes to return.}
}
\value{
A data frame.
}
\description{
A wrapper around \code{\link[snomedizer:concept_ancestors]{snomedizer::concept_descendants()}}.
}
\examples{
\dontrun{
# get children codes for 'Asthma'
 child_codes_sct(
   conceptIds = "195967001",
   include_self = TRUE,
   limit = 1000,
   standardise_output = TRUE
 )
}
}
