% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{code_descriptions_like}
\alias{code_descriptions_like}
\title{Search for codes that match a description}
\usage{
code_descriptions_like(
  reg_expr,
  code_type,
  all_lkps_maps = "all_lkps_maps.db",
  ignore_case = TRUE,
  codes_only = FALSE,
  preferred_description_only = TRUE,
  standardise_output = TRUE
)
}
\arguments{
\item{reg_expr}{a regular expression to search for}

\item{code_type}{character. The type of clinical code system to be searched.
Must be one of \code{read2}, \code{read3}, \code{icd9}, \code{icd10},
\code{bnf}, \code{dmd}, \code{read2_drugs} or \code{opcs4}.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}

\item{ignore_case}{Should case differences be ignored in the match?}

\item{codes_only}{bool. If \code{TRUE}, return a character vector of
\emph{unique} codes. If \code{FALSE} (default), return a data frame of all
results including code descriptions (useful for manual validation).}

\item{preferred_description_only}{bool. Return only preferred descriptions
for clinical codes with synonyms. Default value is \code{TRUE}.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns for the relevant lookup sheet from
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank
resource 592}).}
}
\value{
data frame by default, or a character vector of codes if
\code{codes_only} is \code{TRUE}.
}
\description{
Returns a data frame with clinical codes that match the supplied regular
expression. Ignores case by default.
}
