% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{code_descriptions_like_sct}
\alias{code_descriptions_like_sct}
\title{Search for SNOMED codes with descriptions matching one or more terms}
\usage{
code_descriptions_like_sct(expr, standardise_output = TRUE, limit = 1000)
}
\arguments{
\item{expr}{A pipe ("|") delimited string of code descriptions to search for.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns for the relevant lookup sheet from
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank
resource 592}).}

\item{limit}{Number of SNOMED codes to return.}
}
\value{
A data frame.
}
\description{
A wrapper around \code{\link[snomedizer:concept_find]{snomedizer::concept_find()}}. Note that this also searches
synonym descriptions. For example, searching for 'asth' picks up
'asthenopia', but this is not immediately apparent without looking up synonyms as the returned primary description is .
}
\examples{
\dontrun{
 code_descriptions_like_sct(
   expr = "asth|copd",
   limit = 1000,
   standardise_output = TRUE
 )
}
}
