% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codemapper-package.R
\docType{package}
\name{codemapper-package}
\alias{codemapper}
\alias{codemapper-package}
\title{codemapper: Functions for mapping between clinical coding systems}
\description{
Functions for mapping between clinical coding systems.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/codemapper/}
  \item \url{https://github.com/rmgpanw/codemapper}
  \item Report bugs at \url{https://github.com/rmgpanw/codemapper/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID})

}
\keyword{internal}
