% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{default_col_filters}
\alias{default_col_filters}
\title{Default filtering parameters for lookup and mapping tables.}
\usage{
default_col_filters()
}
\value{
A named list.
}
\description{
To be used as \code{col_filters} argument in 'Clinical code lookups and mappings'
functions. Returns a named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where the
names refer to column names in the corresponding table, and the items are
vectors of values to filter for.
}
\examples{
default_col_filters()
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{get_attributes_sct}()},
\code{\link{get_child_codes}()},
\code{\link{get_children_sct}()},
\code{\link{get_mapping_df}()},
\code{\link{get_parents_sct}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
