% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_all_lkps_maps_path}
\alias{dummy_all_lkps_maps_path}
\title{Dummy UK Biobank resource 592 file path}
\usage{
dummy_all_lkps_maps_path()
}
\value{
A string.
}
\description{
Returns the file path to a dummy \href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank resource 592} excel spreadsheet.
}
\seealso{
Other Dummy data: 
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_dummy_all_lkps_maps}()},
\code{\link{read_dummy_ukb_codings}()}
}
\concept{Dummy data}
