% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_caliber_dir_path}
\alias{dummy_caliber_dir_path}
\title{Dummy CALIBER repository}
\usage{
dummy_caliber_dir_path()
}
\value{
A string.
}
\description{
Returns the file path to a dummy data
\href{https://github.com/spiros/chronological-map-phenotypes}{CALIBER} repository.
}
\seealso{
Other Dummy data: 
\code{\link{build_dummy_all_lkps_maps}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_dummy_all_lkps_maps}()},
\code{\link{read_dummy_ukb_codings}()}
}
\concept{Dummy data}
