% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_phecode_lkp_path}
\alias{dummy_phecode_lkp_path}
\title{Dummy Phecode definitions file path}
\usage{
dummy_phecode_lkp_path()
}
\value{
A string.
}
\description{
Returns the file path to a dummy Phecode definitions 1.2 csv file (full
version may be downloaded from
\href{https://phewascatalog.org/phecodes_icd10}{phewascatalog.org}).
}
\examples{
dummy_phecode_lkp_path()
}
\seealso{
Other Dummy data: 
\code{\link{build_all_lkps_maps_dummy}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_caliber_dir_path}()},
\code{\link{dummy_clinical_events_tidy}()},
\code{\link{dummy_icd10_phecode_map_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_all_lkps_maps_dummy}()},
\code{\link{read_icd10_phecode_map_dummy}()},
\code{\link{read_phecode_lkp_dummy}()},
\code{\link{read_ukb_codings_dummy}()}
}
\concept{Dummy data}
