% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_all_lkps_maps_db_metadata}
\alias{get_all_lkps_maps_db_metadata}
\title{Get metadata from \code{all_lkps_maps} database}
\usage{
get_all_lkps_maps_db_metadata(all_lkps_maps = NULL)
}
\arguments{
\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}
}
\value{
A data frame.
}
\description{
Retrieves metadata for all tables as a single data frame.
}
\examples{
\dontrun{
get_all_lkps_maps_db_metadata()
}
}
