% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_child_codes}
\alias{get_child_codes}
\title{Get descendents for a code}
\usage{
get_child_codes(
  codes,
  code_type,
  all_lkps_maps = NULL,
  codes_only = FALSE,
  preferred_description_only = TRUE,
  standardise_output = TRUE,
  col_filters = default_col_filters()
)
}
\arguments{
\item{codes}{character. A vector of code strings to retrieve child codes for.}

\item{code_type}{character. Type of clinical code system to be searched. One
of bnf, dmd, icd9, icd10, read2, read2_drugs, read3, opcs4, data_coding_3, data_coding_4, data_coding_5, data_coding_6, sct, or phecode.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{codes_only}{bool. If \code{TRUE}, return a character vector of
\emph{unique} codes. If \code{FALSE} (default), return a data frame of all
results including code descriptions (useful for manual validation).}

\item{preferred_description_only}{If \code{TRUE} (default), return only preferred
descriptions for clinical codes with synonyms. Will only apply if
\code{standardise_output} is also \code{TRUE}.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns from the relevant look up table.}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\value{
A data frame
}
\description{
Retrieves children codes for a given set of codes (including the codes
themselves). Note that currently it is not possible to retrieve children
codes for certain clinical coding systems, such as Read 3.
}
\examples{
# TODO
}
\seealso{
\code{\link[=get_children_sct]{get_children_sct()}}

Other Clinical code lookups and mappings: 
\code{\link{default_col_filters}()},
\code{\link{get_children_sct}()},
\code{\link{get_mapping_df}()},
\code{\link{get_relatives_sct}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
