% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_children_sct}
\alias{get_children_sct}
\title{Get children for SNOMED codes}
\usage{
get_children_sct(
  codes = "269823000",
  standarise_output = TRUE,
  active_only = FALSE,
  include_self = TRUE,
  include_descendants = TRUE,
  all_lkps_maps = NULL
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes.}

\item{standarise_output}{If \code{TRUE} (default) return a data frame with columns
'code', 'description' and 'code_type'.}

\item{active_only}{If \code{FALSE} (default) return all codes including currently
those which are currently inactive.}

\item{include_self}{If \code{TRUE} (default) include input codes in the result.}

\item{include_descendants}{If \code{TRUE} (default) return all descendant codes,
as well as immediate children.}
}
\value{
A dataframe
}
\description{
Get children for SNOMED codes
}
