% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_icd10_code_alt_code_x_map}
\alias{get_icd10_code_alt_code_x_map}
\title{Get a mapping table for ICD10 codes in ALT_CODE format, with and without 'X'
appended for undivided 3 character codes}
\usage{
get_icd10_code_alt_code_x_map(icd10_lkp)
}
\arguments{
\item{icd10_lkp}{The lookup table for ICD10 codes, \code{icd10_lkp} .}
}
\value{
A data frame with columns \code{ALT_CODE} and \code{ALT_CODE_minus_x}
}
\description{
Get a mapping table for ICD10 codes in ALT_CODE format, with and without 'X'
appended for undivided 3 character codes
}
