% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_icd10_code_range}
\alias{get_icd10_code_range}
\title{Get a vector of ICD10 codes in ALT_CODE format for a specified start/end
range of ICD10 codes}
\usage{
get_icd10_code_range(start_icd10_code, end_icd10_code, icd10_lkp)
}
\arguments{
\item{start_icd10_code}{String}

\item{end_icd10_code}{String}

\item{icd10_lkp}{The ICD10 lookup table. Must have a \code{.rowid} column.}
}
\value{
A character vector of
}
\description{
Note that \code{start_icd10_code} and \code{end_icd10_code} must be of the same length,
unless one ends with 'X'. For example, expanding the range 'A80-A81' is
equivalent to expanding both 'A800-A809' and 'A810-A819'.
}
