% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_mapping_df}
\alias{get_mapping_df}
\title{Get a 'from-to' mapping data frame}
\usage{
get_mapping_df(from, to, all_lkps_maps = "all_lkps_maps.db")
}
\arguments{
\item{from}{A clinical coding system to map from.}

\item{to}{A clinical coding system to map to.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}
}
\value{
A data frame with column names 'from' and 'to'.
}
\description{
Returns a data frame with 'from' and 'to' columns for a specified pair of
coding systems.
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
