% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_phecode_definitions}
\alias{get_phecode_definitions}
\title{Download the Phecode 1.2 definitions file}
\usage{
get_phecode_definitions()
}
\value{
File path to downloaded file in \code{tempdir()}.
}
\description{
Download link obtained from https://phewascatalog.org/phecodes.
}
