% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_phecode_icd10_map}
\alias{get_phecode_icd10_map}
\title{Download the Phecode 1.2 to ICD10 (beta) mapping file}
\usage{
get_phecode_icd10_map(
  path = file.path(tempdir(), "Phecode_map_v1_2_icd10_beta.csv.zip")
)
}
\arguments{
\item{path}{Path where file will be downloaded to.}
}
\value{
File path to downloaded file.
}
\description{
Download link obtained from https://phewascatalog.org/phecodes.
}
\examples{
\dontrun{
get_phecode_icd10_map()
}
}
