% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_phecode_icd9_map}
\alias{get_phecode_icd9_map}
\title{Download the Phecode 1.2 to ICD9 mapping file}
\usage{
get_phecode_icd9_map(
  path = file.path(tempdir(), "phecode_icd9_map_unrolled.csv.zip")
)
}
\arguments{
\item{path}{Path where file will be downloaded to.}
}
\value{
File path to downloaded file.
}
\description{
Download link obtained from https://phewascatalog.org/phecodes.
}
\examples{
\dontrun{
get_phecode_icd9_map()
}
}
