% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_relatives_sct}
\alias{get_relatives_sct}
\title{Get related SNOMED codes}
\usage{
get_relatives_sct(
  codes,
  relationship = NULL,
  relationship_direction = "child",
  recursive = TRUE,
  active_only = FALSE,
  all_lkps_maps = NULL
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes}

\item{relationship}{SNOMED code defining the type of relationship. By default
this is 'Is a (attribute)'.}

\item{relationship_direction}{Either 'child' (default) or 'parent'.}

\item{recursive}{If \code{TRUE} (default), will recursively search for related
codes e.g. find all descendants code instead of just the immediate child
codes.}

\item{active_only}{If \code{FALSE} (default), return all relationships, even if
currently inactive.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}
}
\value{
A data frame
}
\description{
Low level function for querying related SNOMED codes.
}
\examples{
# get children codes for "269823000" ("Hemoglobin A1C - ...")
get_relatives_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "child"
  ) \%>\%
  lookup_codes("sct")

# get parent codes
get_relatives_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "parent"
  ) \%>\%
  lookup_codes("sct")

# get all codes for which "386868003" ("Bisoprolol (substance)") is an attribute.
# Note, includes "293967003" ("Allergy to bisoprolol (finding)")
get_relatives_sct(
  codes = "386868003",
  relationship = NULL,
  recursive = FALSE,
  relationship_direction = "child"
  ) \%>\%
  lookup_codes("sct")

# more results (i.e. bisoprolol-containing medications) are returned with `recursive = TRUE`
get_relatives_sct(
  codes = "386868003",
  relationship = NULL,
  recursive = FALSE,
  relationship_direction = "child"
  ) \%>\%
  lookup_codes("sct")
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{default_col_filters}()},
\code{\link{get_child_codes}()},
\code{\link{get_children_sct}()},
\code{\link{get_mapping_df}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
