% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_ukb_all_lkps_maps}
\alias{get_ukb_all_lkps_maps}
\title{Get UK Biobank resource 592 directly from UKB
website}
\usage{
get_ukb_all_lkps_maps(dir_path = tempdir())
}
\arguments{
\item{dir_path}{Directory path to download to.}
}
\value{
File path to downloaded \code{all_lkps_maps_v3.xlsx}.
}
\description{
Downloads the UK Biobank code mappings file (\code{all_lkps_maps_v3.xlsx},
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592})
directly from the UKB website.
}
\details{
\strong{Note:} This is a large object (>450 MB)
}
