% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_ukb_self_report_med_to_atc_map}
\alias{get_ukb_self_report_med_to_atc_map}
\title{Download and read a UKB welf-reported medication code to ATC mapping file}
\usage{
get_ukb_self_report_med_to_atc_map()
}
\value{
A data frame.
}
\description{
Mapping table obtained from \href{https://www.nature.com/articles/s41467-019-09572-5#Sec23}{Wray et al 2019},
Supplementary Data 1.
}
