% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{HAS_ATTRIBUTES}
\alias{HAS_ATTRIBUTES}
\title{Get SNOMED codes with a specific set of one or more attributes}
\usage{
HAS_ATTRIBUTES(
  attribute_codes,
  relationship_type = NULL,
  standardise_output = TRUE,
  all_lkps_maps = NULL,
  preferred_description_only = TRUE,
  col_filters = getOption("codemapper.col_filters")
)
}
\arguments{
\item{attribute_codes}{Character vector of SNOMED codes.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns from the relevant look up table.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{preferred_description_only}{If \code{TRUE} (default), return only preferred
descriptions for clinical codes with synonyms. Will only apply if
\code{standardise_output} is also \code{TRUE}.}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\value{
A dataframe
}
\description{
Optionally filtered for specific relationship types. See examples (credit to
\href{https://snomedizer.web.app/articles/snomedizer.html}{snomedizer}).
}
\examples{
# Conditions associated with multiple sclerosis
HAS_ATTRIBUTES("24700007", relationship_type = "42752001")

# Medicines with active ingredient timolol maleate
HAS_ATTRIBUTES("75359005", relationship_type = "10362801000001104")

# Medicines with active ingredient beta blocker
HAS_ATTRIBUTES(CHILDREN("373254001", code_type = "sct"), relationship_type = "10362801000001104")

# Conditions that are caused by bacteria belonging to Enterobacteriaceae
HAS_ATTRIBUTES(CHILDREN("106544002", code_type = "sct"), relationship_type = "246075003")

# Infectious conditions that are caused by bacteria belonging to Enterobacteriaceae
HAS_ATTRIBUTES(CHILDREN("106544002", code_type = "sct"), relationship_type = "246075003") \%AND\% CHILDREN("40733004", code_type = "sct")
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{GET_ATTRIBUTES}()},
\code{\link{default_col_filters}()},
\code{\link{get_attributes_sct}()},
\code{\link{get_child_codes}()},
\code{\link{get_children_sct}()},
\code{\link{get_mapping_df}()},
\code{\link{get_parents_sct}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
