% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infix_setops}
\alias{infix_setops}
\alias{\%AND\%}
\alias{\%OR\%}
\alias{\%NOT\%}
\title{Infix set operations}
\usage{
x \%AND\% y

x \%OR\% y

x \%NOT\% y
}
\arguments{
\item{x, y}{Pair of compatible data frames. A pair of data frames is
compatible if they have the same column names (possibly in different
orders) and compatible types.}
}
\description{
Infix versions of \link[dplyr:setops]{dplyr::setops}. \verb{\%AND\%}, \verb{\%OR\%} and \verb{\%NOT\%} are the
equivalents of \code{\link[dplyr:setops]{dplyr::intersect()}}, \code{\link[dplyr:setops]{dplyr::union()}} and \code{\link[dplyr:setops]{dplyr::setdiff()}}
respectively.
}
\examples{
df1 <- tibble::tibble(x = 1:3)
df2 <- tibble::tibble(x = 3:5)

# equivalent of dplyr::intersect(df1, df2)
df1 \%AND\% df2

# equivalent of dplyr::union(df1, df2)
df1 \%OR\% df2

# equivalent of dplyr::setdiff(df1, df2). Note that argument order matters
df1 \%NOT\% df2
df2 \%NOT\% df1
}
