% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{lookup_codes}
\alias{lookup_codes}
\title{Look up descriptions for clinical codes}
\usage{
lookup_codes(
  codes,
  code_type,
  all_lkps_maps = NULL,
  preferred_description_only = TRUE,
  standardise_output = TRUE,
  unrecognised_codes = "error",
  col_filters = default_col_filters(),
  .return_unrecognised_codes = FALSE
)
}
\arguments{
\item{codes}{character. Vector of codes to lookup}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{preferred_description_only}{bool. Return only preferred descriptions
for clinical codes with synonyms. Default value is \code{TRUE}.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns from the relevant look up table.}

\item{unrecognised_codes}{Either 'error' (default) or 'warning'. If any input
\code{codes} are unrecognised, then either an error or warning will be raised.}

\item{.return_unrecognised_codes}{If \code{TRUE}, return a vector of unrecognised
codes only.}

\item{codes_only}{bool. If \code{TRUE}, return a character vector of
\emph{unique} codes. If \code{FALSE} (default), return a data frame of all
results including code descriptions (useful for manual validation).}
}
\value{
data frame
}
\description{
Returns a data frame including descriptions for the codes of interest
}
\examples{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# look up ICD10 codes
lookup_codes(
  codes = c("E10", "E11"),
  code_type = "icd10",
  all_lkps_maps = all_lkps_maps_dummy
)
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{default_col_filters}()},
\code{\link{get_mapping_df}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
