% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_events_to_phecodes.R
\name{map_clinical_events_to_phecodes}
\alias{map_clinical_events_to_phecodes}
\title{Map UK Biobank clinical events to phecodes}
\usage{
map_clinical_events_to_phecodes(
  clinical_events,
  all_lkps_maps = NULL,
  min_date_only = FALSE,
  col_filters = default_col_filters()
)
}
\arguments{
\item{clinical_events}{A long format data frame created by
\code{\link[ukbwranglr]{tidy_clinical_events}}, \code{\link[ukbwranglr]{tidy_gp_clinical}},
\code{\link[ukbwranglr]{tidy_gp_scripts}} or \code{\link[ukbwranglr]{make_clinical_events_db}}.
This can also be a \code{\link[dbplyr]{tbl_dbi}} object.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an SQLite
database named 'all_lkps_maps.db' in the current working directory, or to a
a SQLite database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{min_date_only}{If \code{TRUE}, result will be filtered for only the earliest
date per eid-phecode pair (date will be recorded as \code{NA} for cases where
there are no dates).}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\value{
A data frame with column names 'eid', 'source', 'index', 'code',
'icd10', 'phecode' and 'date'.
}
\description{
UK Biobank clinical events sources that are recorded in ICD10 are mapped
directly to phecodes, while non-ICD10 sources are mapped to phecodes via
ICD10.
}
\details{
Maps the following UK Biobank clinical events sources to phecodes: f40001, f40002, f20002_icd10, f40006, f41270, f40013, f41271, gpc1_r3, gpc2_r3, gpc3_r3, gpc4_r3, gpc1_r2, gpc2_r2, gpc3_r2, gpc4_r2.
}
\examples{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# dummy clinical events data frame
dummy_clinical_events_tidy()

# map to phecodes
map_clinical_events_to_phecodes(
  clinical_events = dummy_clinical_events_tidy(),
  all_lkps_maps = all_lkps_maps_dummy,
  min_date_only = FALSE
)
}
