% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{map_codes}
\alias{map_codes}
\title{Map clinical codes from one coding system to another}
\usage{
map_codes(
  codes,
  from,
  to,
  all_lkps_maps = NULL,
  codes_only = FALSE,
  standardise_output = TRUE,
  unrecognised_codes = "error",
  preferred_description_only = NULL,
  reverse_mapping = "error",
  col_filters = default_col_filters()
)
}
\arguments{
\item{codes}{A character vector of codes to be mapped.}

\item{from}{Coding system that \code{codes} belong to.}

\item{to}{Coding system to map \code{codes} to.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an SQLite
database named 'all_lkps_maps.db' in the current working directory, or to a
a SQLite database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{codes_only}{bool. If \code{TRUE}, return a character vector of
\emph{unique} codes. If \code{FALSE} (default), return a data frame of all
results including code descriptions (useful for manual validation).}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns for the relevant lookup sheet from
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank
resource 592}).}

\item{unrecognised_codes}{Either 'error' (default) or 'warning'. If any input
\code{codes} are unrecognised for the coding system being mapped from, then
either an error or warning will be raised.}

\item{preferred_description_only}{If \code{TRUE}, return only preferred
descriptions for clinical codes with synonyms. Can only be \code{TRUE} if
\code{standardise_output} is also \code{TRUE}. Default value is
\code{NULL}.}

\item{reverse_mapping}{If 'error' (default), an error raised if attempting to
map between coding systems for which a mapping table does not exist. If
'warning', will raise a warning and attempt to use an existing mapping
table in the opposite direction (for example, a mapping from ICD10 to Read
3 would be attempted using the Read 3-to-ICD10 mapping table).}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\description{
Uses the code mapping file provided by UK Biobank
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592}).
}
\details{
The values for arguments \code{from} and \code{to} must be one of
\code{read2}, \code{read3}, \code{icd9}, \code{icd10}, \code{bnf},
\code{dmd}, \code{read2_drugs} or \code{opcs4}.
}
\examples{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# map codes from Read 2 to ICD10
map_codes(codes = "G20..",
         from = "read2",
         to = "icd10",
         all_lkps_maps = all_lkps_maps_dummy)
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{default_col_filters}()},
\code{\link{get_mapping_df}()},
\code{\link{lookup_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
