% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecodes.R
\name{map_codes_ukb_clinical_events}
\alias{map_codes_ukb_clinical_events}
\title{Map codes in a UKB clinical events table}
\usage{
map_codes_ukb_clinical_events(
  clinical_events,
  from,
  to,
  all_lkps_maps = "all_lkps_maps.db",
  strict_ukb = TRUE
)
}
\arguments{
\item{clinical_events}{A clinical events table created by
\code{\link[ukbwranglr:tidy_clinical_events]{ukbwranglr::tidy_clinical_events()}}}

\item{from}{Coding type being mapped from}

\item{to}{Coding type being mapped to}

\item{all_lkps_maps}{Named list of SQLite database with lookup/mapping tables}

\item{strict_ukb}{If \code{TRUE}, extra checks are performed to check that
\code{source} and \code{data_coding} match \code{\link[ukbwranglr:clinical_events_sources]{ukbwranglr::clinical_events_sources()}}.}
}
\value{
A clinical events dataframe
}
\description{
Joins the existing \code{code} column to a mapping data frame created by
\link{get_mapping_df}, then filters out any missing values for the new
code type, then replaces the old \code{code} column with the newly mapped codes.
}
\details{
Clinical events can contain multiple sources, but these should all use the
same data_coding. An error is raised otherwise.
}
