% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{read_all_lkps_maps}
\alias{read_all_lkps_maps}
\title{Read UK Biobank resource 592 into a named list}
\usage{
read_all_lkps_maps(path = get_ukb_all_lkps_maps())
}
\arguments{
\item{path}{Path to \code{all_lkps_maps_v4.xlsx}. By default, this is downloaded
from the UK Biobank website using \code{\link[=get_ukb_all_lkps_maps]{get_ukb_all_lkps_maps()}}.}
}
\value{
A named list of data frames.
}
\description{
Reads the UK Biobank code mappings file (\code{all_lkps_maps_v4.xlsx},
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{resource 592}) into
a named list of data frames.
}
\details{
\strong{Note:} This is a large object (>450 MB)
}
\examples{
read_all_lkps_maps(dummy_all_lkps_maps_path())
}
