% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{read_all_lkps_maps_dummy}
\alias{read_all_lkps_maps_dummy}
\title{Read dummy UK Biobank resource 592 into R}
\usage{
read_all_lkps_maps_dummy()
}
\value{
A named list of tibbles.
}
\description{
Reads a dummy \href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank resource 592} excel spreadsheet
into R.
}
\examples{
read_all_lkps_maps_dummy()
}
\seealso{
Other Dummy data: 
\code{\link{build_all_lkps_maps_dummy}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_caliber_dir_path}()},
\code{\link{dummy_clinical_events_tidy}()},
\code{\link{dummy_phecode_lkp_path}()},
\code{\link{dummy_pheode_icd10_map_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_phecode_icd10_map_dummy}()},
\code{\link{read_phecode_lkp_dummy}()},
\code{\link{read_ukb_codings_dummy}()}
}
\concept{Dummy data}
