% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{read_dummy_all_lkps_maps}
\alias{read_dummy_all_lkps_maps}
\title{Read dummy UK Biobank resource 592 into R}
\usage{
read_dummy_all_lkps_maps()
}
\value{
A named list of tibbles.
}
\description{
Reads a dummy \href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank resource 592} excel spreadsheet
into R.
}
\examples{
read_dummy_all_lkps_maps()
}
\seealso{
Other Dummy data: 
\code{\link{build_dummy_all_lkps_maps}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_dummy_ukb_codings}()}
}
\concept{Dummy data}
