% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{read_dummy_ukb_codings}
\alias{read_dummy_ukb_codings}
\title{Read dummy UK Biobank codings into R}
\usage{
read_dummy_ukb_codings()
}
\value{
A data frame.
}
\description{
Reads a dummy \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UK Biobank codings}
tsv file into R.
}
\seealso{
Other Dummy data: 
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_dummy_all_lkps_maps}()}
}
\concept{Dummy data}
