% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{read_phecode_lkp_dummy}
\alias{read_phecode_lkp_dummy}
\title{Read dummy Phecode definitions file into R}
\usage{
read_phecode_lkp_dummy()
}
\value{
A data frame.
}
\description{
Reads a dummy Phecode definitions 1.2 csv file into R (full version may be
downloaded from
\href{https://phewascatalog.org/phecodes_icd10}{phewascatalog.org})
}
\examples{
read_phecode_lkp_dummy()
}
\seealso{
Other Dummy data: 
\code{\link{build_all_lkps_maps_dummy}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_caliber_dir_path}()},
\code{\link{dummy_clinical_events_tidy}()},
\code{\link{dummy_phecode_lkp_path}()},
\code{\link{dummy_pheode_icd10_map_path}()},
\code{\link{dummy_ukb_codings_path}()},
\code{\link{read_all_lkps_maps_dummy}()},
\code{\link{read_phecode_icd10_map_dummy}()},
\code{\link{read_ukb_codings_dummy}()}
}
\concept{Dummy data}
