% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{reformat_icd10_codes}
\alias{reformat_icd10_codes}
\title{Reformat ICD-10 codes}
\usage{
reformat_icd10_codes(
  icd10_codes,
  all_lkps_maps = NULL,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE",
  unrecognised_codes = "error",
  strip_x = FALSE
)
}
\arguments{
\item{icd10_codes}{character vector of ICD-10 codes}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an SQLite
database named 'all_lkps_maps.db' in the current working directory, or to a
a SQLite database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{input_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}

\item{output_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}

\item{unrecognised_codes}{Either 'error' (default) or 'warning'. If any input
\code{codes} are unrecognised, then either an error or warning will be raised.}

\item{strip_x}{If \code{TRUE} and converting to \code{ALT_CODE} format, 'X' is removed
from the end of undivided 3 character codes (default is \code{FALSE}).}
}
\value{
character vector of ICD-10 codes, reformatted as specified by
\code{output_icd10_format}.
}
\description{
The lookup sheet in
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UKB resource 592}
for ICD-10 ("icd10_lkp") has a column called \code{ALT_CODE}, which is an
alternative format for ICD-10 codes. This is the format used in the mapping
sheets for this resource, as well as in
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=41270}{Field ID 41270}.
\emph{Note however that in contrast to the ICD10 lookup and mapping sheets,
undivided 3 character codes do not have an 'X' appended in the UK Biobank
dataset}. This function converts from one format to the other, with the
option to strip the final 'X' from undivided 3 character codes if converting
to \code{ALT_CODE} format.
}
\examples{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# reformat from ICD10_CODE to ALT_CODE
reformat_icd10_codes(icd10_codes = c("E10.9"),
  all_lkps_maps = all_lkps_maps_dummy,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE")

# reformat from ALT_CODE to ICD10_CODE
reformat_icd10_codes(icd10_codes = c("E109"),
  all_lkps_maps = all_lkps_maps_dummy,
  input_icd10_format = "ALT_CODE",
  output_icd10_format = "ICD10_CODE")
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{default_col_filters}()},
\code{\link{get_mapping_df}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()}
}
\concept{Clinical code lookups and mappings}
