% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_codelistBuilder.R
\name{RunCodelistBuilder}
\alias{RunCodelistBuilder}
\title{Shiny app for building clinical codelists.}
\usage{
RunCodelistBuilder(
  all_lkps_maps = NULL,
  options = list(launch.browser = TRUE),
  ...
)
}
\arguments{
\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{...}{Additional args passed on to \code{\link[shiny]{shinyApp}}}
}
\value{
\code{NULL}
}
\description{
Launches a Shiny app. Note, requires \code{all_lkps_maps} database file.
}
\examples{
\dontrun{
# build dummy all_lkps_maps database
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()
all_lkps_maps_db <- all_lkps_maps_to_db(
  all_lkps_maps = all_lkps_maps_dummy,
  db_path = tempfile(fileext = ".db")
)

# launch app
RunCodelistBuilder(all_lkps_maps = all_lkps_maps_db)
}
}
