% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCodeMapper}
\alias{runCodeMapper}
\title{Launch a shiny app to select clinical codes}
\usage{
runCodeMapper(all_lkps_maps, options = list(launch.browser = TRUE), ...)
}
\arguments{
\item{all_lkps_maps}{A named list or path to SQLite database}

\item{...}{Additional args passed on to \code{\link[shiny]{shinyApp}}}
}
\value{
\code{NULL}
}
\description{
TODO
}
