% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{build_all_lkps_maps}
\alias{build_all_lkps_maps}
\title{Build named list of clinical code look up and mapping tables}
\usage{
build_all_lkps_maps(
  all_lkps_maps = read_all_lkps_maps(),
  ukb_codings = ukbwranglr::get_ukb_codings_direct(),
  bnf_dmd = get_nhsbsa_snomed_bnf(),
  self_report_med_to_atc_map = get_ukb_self_report_med_to_atc_map(),
  ctv3sctmap2 = NULL,
  phecode_1_2_lkp = NULL,
  icd10_phecode_1_2 = NULL,
  icd9_phecode_1_2 = NULL
)
}
\arguments{
\item{all_lkps_maps}{UK Biobank resource 592, as returned by
\code{\link{get_ukb_all_lkps_maps}}.}

\item{ukb_codings}{The UK Biobank codings file, as returned by
\code{\link[ukbwranglr]{get_ukb_codings_direct}}.}

\item{bnf_dmd}{Optional: a data frame, the NHSBSA BNF-SNOMED mapping table as
returned by \code{\link{get_nhsbsa_snomed_bnf}}.}

\item{self_report_med_to_atc_map}{Optional: a data frame (see
\code{\link[=get_ukb_self_report_med_to_atc_map]{get_ukb_self_report_med_to_atc_map()}}).}

\item{ctv3sctmap2}{Optional: path to the NHS TRUD mapping file for Read 3 to
SNOMEDCT ("ctv3sctmap2_uk_20200401000001.txt").}

\item{phecode_1_2_lkp}{Optional: path to the phecode v1.2 lookup file
("phecode_definitions1.2.csv.zip").}

\item{icd10_phecode_1_2}{Optional: path to the phecode v1.2 to ICD10 mapping
file ("Phecode_map_v1_2_icd10_beta.csv.zip").}

\item{icd9_phecode_1_2}{Optional: path to the phecode v1.2 to ICD10 mapping
file ("phecode_icd9_map_unrolled.csv.zip").}
}
\value{
Returns a named list of data frames.
}
\description{
Downloads the lookup and mapping tables from
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank resource
592} as well as the NHSBSA BNF-SNOMED mapping table (available
\href{https://www.nhsbsa.nhs.uk/prescription-data/understanding-our-data/bnf-snomed-mapping}{here})
and OPCS4, self-reported medical conditions/medications/operations from the
UK Biobank codings file (available
\href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{here}).
}
\seealso{
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}
}
