% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{codes_starting_with}
\alias{codes_starting_with}
\title{Get codes that start with...}
\usage{
codes_starting_with(
  codes,
  code_type,
  all_lkps_maps = NULL,
  codes_only = FALSE,
  preferred_description_only = TRUE,
  standardise_output = TRUE,
  col_filters = default_col_filters(),
  limit = 1000
)
}
\arguments{
\item{codes}{character. A vector of code strings to search for matching
codes.}

\item{code_type}{character. The type of clinical code system to be searched.
Must be one of bnf, dmd, icd9, icd10, read2, read2_drugs, read3, opcs4, data_coding_3, data_coding_4, data_coding_5, data_coding_6, sct, or phecode.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an SQLite
database named 'all_lkps_maps.db' in the current working directory, or to a
a SQLite database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{codes_only}{bool. If \code{TRUE}, return a character vector of
\emph{unique} codes. If \code{FALSE} (default), return a data frame of all
results including code descriptions (useful for manual validation).}

\item{preferred_description_only}{bool. Return only preferred descriptions
for clinical codes with synonyms. Default value is \code{TRUE}.}

\item{standardise_output}{bool. If \code{TRUE} (default), outputs a data
frame with columns named 'code', 'description' and 'code_type'. Otherwise
returns a data frame with all columns for the relevant lookup sheet from
(\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UK Biobank
resource 592}).}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}

\item{limit}{Integer. Only applies to SNOMED codes (passed on to relevant
\link[snomedizer:snomedizer-package]{snomedizer::snomedizer} function).}
}
\description{
This is case \emph{sensitive} (important for read codes especially).
}
\examples{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# lookup codes
lookup_codes(
  codes = c("E10", "E11"),
  code_type = "icd10",
  all_lkps_maps = all_lkps_maps_dummy
)
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{default_col_filters}()},
\code{\link{get_mapping_df}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
