% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy_ukb_codings_path}
\alias{dummy_ukb_codings_path}
\title{Dummy UK Biobank codings file path}
\usage{
dummy_ukb_codings_path()
}
\value{
A string.
}
\description{
Returns the file path to a dummy \href{https://biobank.ctsu.ox.ac.uk/crystal/exinfo.cgi?src=accessing_data_guide}{UK Biobank codings}
tsv file.
}
\examples{
dummy_ukb_codings_path()
}
\seealso{
Other Dummy data: 
\code{\link{build_dummy_all_lkps_maps}()},
\code{\link{dummy_all_lkps_maps_path}()},
\code{\link{dummy_caliber_dir_path}()},
\code{\link{read_dummy_all_lkps_maps}()},
\code{\link{read_dummy_ukb_codings}()}
}
\concept{Dummy data}
