% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_children_sct}
\alias{get_children_sct}
\title{Get children for SNOMED codes}
\usage{
get_children_sct(
  codes = "269823000",
  standarise_output = TRUE,
  active_only = FALSE,
  include_self = TRUE,
  include_descendants = TRUE,
  all_lkps_maps = NULL
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes.}

\item{standarise_output}{If \code{TRUE} (default) return a data frame with columns
'code', 'description' and 'code_type'.}

\item{active_only}{If \code{FALSE} (default) return all codes including those
which are currently inactive.}

\item{include_self}{If \code{TRUE} (default) include input codes in the result.}

\item{include_descendants}{If \code{TRUE} (default) return all descendant codes,
as well as immediate children.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}
}
\value{
A dataframe
}
\description{
Get children for SNOMED codes
}
