% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_children_sct}
\alias{get_children_sct}
\title{Get children for SNOMED codes}
\usage{
get_children_sct(
  codes,
  standardise_output = TRUE,
  include_self = TRUE,
  include_descendants = TRUE,
  all_lkps_maps = NULL,
  preferred_description_only = TRUE,
  col_filters = getOption("codemapper.col_filters")
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes.}

\item{standardise_output}{If \code{TRUE} (default) return a data frame with columns
'code', 'description' and 'code_type'.}

\item{include_self}{If \code{TRUE} (default) include input codes in the result.}

\item{include_descendants}{If \code{TRUE} (default) return all descendant codes,
as well as immediate children.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{preferred_description_only}{If \code{TRUE} (default), return only preferred
descriptions for clinical codes with synonyms. Will only apply if
\code{standardise_output} is also \code{TRUE}.}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\value{
A dataframe
}
\description{
Get children for SNOMED codes
}
\seealso{
\code{\link[=get_child_codes]{get_child_codes()}}, \code{\link[=get_parents_sct]{get_parents_sct()}}

Other Clinical code lookups and mappings: 
\code{\link{GET_ATTRIBUTES}()},
\code{\link{HAS_ATTRIBUTES}()},
\code{\link{default_col_filters}()},
\code{\link{get_attributes_sct}()},
\code{\link{get_child_codes}()},
\code{\link{get_mapping_df}()},
\code{\link{get_parents_sct}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
