% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_mapping_df}
\alias{get_mapping_df}
\title{Get a 'from-to' mapping data frame}
\usage{
get_mapping_df(
  from,
  to,
  all_lkps_maps = "all_lkps_maps.db",
  rename_from_to = NULL,
  na.rm = TRUE,
  reverse_mapping = "error",
  col_filters = default_col_filters()
)
}
\arguments{
\item{from}{A clinical coding system to map from.}

\item{to}{A clinical coding system to map to.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}

\item{rename_from_to}{Optionally supply a named vector to rename the 'from'
and 'to' columns. For example \code{c(from = "original_codes", to = "new_codes")}. By default, the columns will be named using the values for
\code{from} and \code{to} arguments.}

\item{na.rm}{If \code{TRUE} (default), remove any rows with \code{NA} from the returned
mapping data frame. The mapping tables may sometimes include \code{NA} values to
explicitly show which 'from' codes have not been mapped.}

\item{reverse_mapping}{If 'error' (default), an error raised if attempting to
map between coding systems for which a mapping table does not exist. If
'warning', will raise a warning and attempt to use an existing mapping
table in the opposite direction (for example, a mapping from ICD10 to Read
3 would be attempted using the Read 3-to-ICD10 mapping table).}

\item{col_filters}{A named list where each name in the list refers to the
name of a lookup or mapping table. Each item is also a named list, where
the names refer to column names in the corresponding table, and the items
are vectors of values to filter for. For example, \verb{list(my_lookup_table = list(colA = c("A", "B"))} will result in \code{my_lookup_table} being filtered
for rows where \code{colA} is either 'A' or 'B'. Uses \code{default_col_filters()} by
default. Set to \code{NULL} to remove all filters.}
}
\value{
A data frame with column names 'from' and 'to'.
}
\description{
Returns a data frame with 'from' and 'to' columns for a specified pair of
coding systems.
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{default_col_filters}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()},
\code{\link{reformat_icd10_codes}()}
}
\concept{Clinical code lookups and mappings}
