% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_nhsbsa_snomed_bnf}
\alias{get_nhsbsa_snomed_bnf}
\title{Download and read the NHSBSA BNF_SNOMED mapping file}
\usage{
get_nhsbsa_snomed_bnf(path = file.path(tempdir(), "bnf_dmd.zip"))
}
\arguments{
\item{path}{Path where file will be downloaded to.}
}
\value{
File path to downloaded file.
}
\description{
Mapping table available from
\href{https://www.nhsbsa.nhs.uk/prescription-data/understanding-our-data/bnf-snomed-mapping}{here}.
}
\examples{
\dontrun{ get_nhsbsa_snomed_bnf() }
}
