% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_phecode_definitions}
\alias{get_phecode_definitions}
\title{Download the Phecode 1.2 definitions file}
\usage{
get_phecode_definitions(
  path = file.path(tempdir(), "phecode_definitions1.2.csv.zip")
)
}
\arguments{
\item{path}{Path where file will be downloaded to.}
}
\value{
File path to downloaded file.
}
\description{
Download link obtained from https://phewascatalog.org/phecodes.
}
\examples{
\dontrun{ get_phecode_definitions() }
}
