% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups_and_mappings.R
\name{get_phecode_icd10_map}
\alias{get_phecode_icd10_map}
\title{Download the Phecode 1.2 to ICD10 (beta) mapping file}
\usage{
get_phecode_icd10_map()
}
\value{
File path to downloaded file in \code{tempdir()}.
}
\description{
Download link obtained from https://phewascatalog.org/phecodes.
}
