% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_relatives_sct}
\alias{get_relatives_sct}
\title{Get related SNOMED codes}
\usage{
get_relatives_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "child",
  recursive = TRUE,
  active_only = FALSE,
  all_lkps_maps = NULL
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes}

\item{relationship}{SNOMED code defining the type of relationship. By default
this is 'Is a (attribute)'.}

\item{relationship_direction}{Either 'child' (default) or 'parent'.}

\item{recursive}{If \code{TRUE} (default), will recursively search for related
codes e.g. find all descendants code instead of just the immediate child
codes.}

\item{active_only}{If \code{FALSE} (default), return all relationships, even if
currently inactive.}
}
\value{
A data frame
}
\description{
Low level function for querying related SNOMED codes.
}
