% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{get_relatives_sct}
\alias{get_relatives_sct}
\title{Get related SNOMED codes}
\usage{
get_relatives_sct(
  codes = "269823000",
  relationship = "116680003",
  relationship_direction = "child",
  recursive = TRUE,
  active_only = FALSE,
  all_lkps_maps = NULL
)
}
\arguments{
\item{codes}{Character vector of SNOMED codes}

\item{relationship}{SNOMED code defining the type of relationship. By default
this is 'Is a (attribute)'.}

\item{relationship_direction}{Either 'child' (default) or 'parent'.}

\item{recursive}{If \code{TRUE} (default), will recursively search for related
codes e.g. find all descendants code instead of just the immediate child
codes.}

\item{active_only}{If \code{FALSE} (default), return all relationships, even if
currently inactive.}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables
(either data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables. If \code{NULL}, will attempt to connect to a Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to
a a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}
}
\value{
A data frame
}
\description{
Low level function for querying related SNOMED codes.
}
