% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_codes.R
\name{reformat_icd10_codes}
\alias{reformat_icd10_codes}
\title{Reformat ICD-10 codes}
\usage{
reformat_icd10_codes(
  icd10_codes,
  all_lkps_maps,
  input_icd10_format = "ICD10_CODE",
  output_icd10_format = "ALT_CODE",
  strip_x = FALSE
)
}
\arguments{
\item{icd10_codes}{character vector of ICD-10 codes}

\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a SQLite database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). By default, will look for a SQLite datbase called
\code{all_lkps_maps.db} in the current working directory.}

\item{input_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}

\item{output_icd10_format}{character. Must be either "ICD10_CODE" or
"ALT_CODE".}

\item{strip_x}{logical. If \code{TRUE}, any 3 character ICD10 codes in
'ALT_CODE' format with no children (e.g. 'A38X', Scarlet fever) will have
the last 'X' removed. Default value is \code{FALSE}.}
}
\value{
character vector of ICD-10 codes, reformatted as specified by
\code{output_icd10_format}.
}
\description{
The lookup sheet in
\href{https://biobank.ndph.ox.ac.uk/ukb/refer.cgi?id=592}{UKB resource 592}
for ICD-10 ("icd10_lkp") has a column called "ALT_CODE", which is an
alternative format for ICD-10 codes. This is the format used in the mapping
sheets for this resource, as well as in
\href{https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=41270}{Field ID 41270}.
This function converts from one format to the other.
}
\seealso{
Other Clinical code lookups and mappings: 
\code{\link{codes_starting_with}()},
\code{\link{get_mapping_df}()},
\code{\link{lookup_codes}()},
\code{\link{map_codes}()}
}
\concept{Clinical code lookups and mappings}
