% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCodelistBuilder.R
\name{RunCodelistBuilder}
\alias{RunCodelistBuilder}
\title{Create a clinical codelist with R Shiny.}
\usage{
RunCodelistBuilder(
  all_lkps_maps = NULL,
  options = list(launch.browser = TRUE),
  ...
)
}
\arguments{
\item{all_lkps_maps}{Either a named list of lookup and mapping tables (either
data frames or \code{tbl_dbi} objects), or the path to a Duckdb database
containing these tables (see also \code{\link[=build_all_lkps_maps]{build_all_lkps_maps()}} and
\code{\link[=all_lkps_maps_to_db]{all_lkps_maps_to_db()}}). If \code{NULL}, will attempt to connect to an Duckdb
database named 'all_lkps_maps.db' in the current working directory, or to a
a Duckdb database specified by an environmental variable named
'ALL_LKPS_MAPS_DB' (see
\href{https://resources.numbat.space/using-rprofile-and-renviron.html#renviron}{here}
for how to set environment variables using a \code{.Renviron} file). The latter
method will be used in preference.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{...}{Additional args passed on to \code{\link[shiny]{shinyApp}}}
}
\value{
\code{NULL}
}
\description{
Launches a Shiny app to select clinical codes (see also 'User notes' section
below). In general: \enumerate{\item Complete details for essential search
fields (leaving 'category' box blank and leaving the 'Ignore case' boxes
ticked - if un-ticked, this will not be recorded under
'description_search_strategy'). \item Optionally upload a previous codelist.
\item Click 'Download selected'.}
}
\section{User notes}{


\itemize{ \item Essential fields: 'Disease', either 'Code descriptions
like...' or at least one of 'Codes starting with...'. \item 'Download all':
downloads all codes under the 'Matching clinical codes' tab,
\emph{disregarding} any selections (i.e. the 'selected' column will appear
blank). \item 'Download selected': downloads all codes under the 'Preview
selected codes only' tab. \item If uploading a previous codelist selection,
this must be formatted as required by
\code{\link[ukbwranglr]{validate_clinical_codes}} (see
\code{\link[ukbwranglr]{example_clinical_codes}} for a valid example).
\item ...this will automatically select any codes matching on 'disease',
'code_type' and 'code' (see \code{update_code_selection} function in
\code{utils.R}), and on clicking 'Download selected', will copy over any
values in the 'category' column and set the 'selected' column to 'Yes' only
for these rows. \item ...the full list of uploaded codes can be viewed
under the 'Uploaded codelist' tab, which includes a column called
'included_in_matching', indicating whether each uploaded code is present
('Yes') or \emph{not} present ('No') in the list under 'Matching clinical
codes'. \item Also, click through the tabs after each step before moving on
to the next. \item The final list of matching codes is codes matching 'Code
descriptions like...' \strong{OR} 'Codes starting with...'. \item
\strong{CAUTION:} make sure there are no inadvertent white spaces (e.g.
'search term', not 'search term  ').}
}

\examples{
\dontrun{
# build dummy all_lkps_maps
all_lkps_maps_dummy <- build_all_lkps_maps_dummy()

# launch app
RunCodelistBuilder(all_lkps_maps = all_lkps_maps_dummy)
}
}
