% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runCodelistBuilder}
\alias{runCodelistBuilder}
\title{Create a clinical codelist.}
\usage{
runCodelistBuilder(all_lkps_maps, options = list(launch.browser = TRUE), ...)
}
\arguments{
\item{all_lkps_maps}{A named list or path to SQLite database of clinical
codes created by \code{\link{build_all_lkps_maps}} and
\code{\link{all_lkps_maps_to_db}}.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{...}{Additional args passed on to \code{\link[shiny]{shinyApp}}}
}
\value{
\code{NULL}
}
\description{
Launches a shiny app to select clinical codes (see also 'User notes' section
below). In general: \enumerate{\item Complete details for essential search
fields (leaving 'category' box blank and leaving the 'Ignore case' boxes
ticked - if un-ticked, this will not be recorded under
'description_search_strategy'). \item Optionally upload a previous codelist.
\item Click 'Download selected'.}
}
\section{User notes}{


\itemize{ \item Essential fields: 'Disease', either 'Code descriptions
like...' or at least one of 'Codes starting with...'. \item 'Download all':
downloads all codes under the 'Matching clinical codes' tab,
\emph{disregarding} any selections (i.e. the 'selected' column will appear
blank). \item 'Download selected': downloads all codes under the 'Preview
selected codes only' tab. \item If uploading a previous codelist selection,
this must be formatted as required by
\code{\link[ukbwranglr]{validate_clinical_codes}} (see
\code{\link[ukbwranglr]{example_clinical_codes}} for a valid example).
\item ...this will automatically select any codes matching on 'disease',
'code_type' and 'code' (see \code{update_code_selection} function in
\code{utils.R}), and on clicking 'Download selected', will copy over any
values in the 'category' column and set the 'selected' column to 'Yes' only
for these rows. \item ...the full list of uploaded codes can be viewed
under the 'Uploaded codelist' tab, which includes a column called
'included_in_matching', indicating whether each uploaded code is present
('Yes') or \emph{not} present ('No') in the list under 'Matching clinical
codes'. \item Also, click through the tabs after each step before moving on
to the next. \item The final list of matching codes is codes matching 'Code
descriptions like...' \strong{OR} 'Codes starting with...'. \item
\strong{CAUTION:} make sure there are no inadvertent white spaces (e.g.
'search term', not 'search term  ').}
}

